+++ Magic/CHANGELOG.md
@@ -253,7 +253,8 @@
  - Add support for getting colored tipped arrows, `/mgive tipped_arrow{18001}:#00FF00`
  - Add `leash_range` mob property, currently only works for NPCs
  - Add pose parameters to armor stands
- 
+ - Add TokenManager integration, adds `tokens` currency 
+
 ## Changes
 
  - Rebranded "automata" to "magic blocks", to avoid confusion with the "block monster" automata